// frontend/src/types/stock.ts
export interface Stock {
  id: string;
  symbol: string;
  company_name: string;
  sector: string;
  exchange: string;
  current_price: string;
  previous_close: string;
  day_high: string;
  day_low: string;
  volume: string;
  market_cap: string;
  pe_ratio: string;
  change: string;
  change_percent: string;
}